refactor(ai): extract transient error classifier to split resilience.ts
resilience.ts ballooned past the 400-line hard cap during the Phoenix
migration. Moved the transient-error classifier into convex/ai/transientErrors.ts
and re-export for existing callers. No behavior change.
